refactored

+++ app/controllers/comments_controller.rb
@@ -2,12 +2,16 @@ class CommentsController < ApplicationController
 
     post '/comments' do 
         
-        user = User.find(session[:user_id])
-        patient = Patient.find(user.patient_id)
+        
         staff = Staff.find(params[:staff][:id]) 
         comment = Comment.create(params[:comment])
-        patient.comments << comment 
-        staff.comments << comment 
+        
+        
+        comment.staff_id = params[:staff][:id]
+        comment.patient_id = current_user.patient_id
+        comment.save 
+        
+        
 
         redirect "staffs/doctors/#{staff.slug}"
     end
@@ -19,7 +23,10 @@ class CommentsController < ApplicationController
 
     patch '/comment/:id' do 
         comment = Comment.find(params[:id])
-        comment.update(content: params[:comment][:content])
+        
+        if logged_in? && comment.patient.user.id == session[:user_id]
+            comment.update(content: params[:comment][:content])
+        end
         staff = Staff.find(comment.staff_id)
         flash[:message] = "Successfully updated a comment!!"
         redirect "/staffs/doctors/#{staff.slug}"
@@ -27,8 +34,11 @@ class CommentsController < ApplicationController
 
     delete '/comment/:id' do 
         comment = Comment.find(params[:id])
-        staff = Staff.find(comment.staff_id)
-        comment.destroy
+        # staff = Staff.find(comment.staff_id)
+        staff = comment.staff 
+        if logged_in? && comment.patient.user.id == session[:user_id]
+            comment.destroy
+        end
         flash[:message] = "Successfully deleted a comment!!"
         redirect "/staffs/doctors/#{staff.slug}"
     end
